refactor: update imports to use shadcn directory in contact, features, plans, profile, and register pages
