+++ app/plans/page.tsx
@@ -6,10 +6,10 @@ import { MarketingNav } from '@/features/home/components/marketing-nav';
 import { MarketingFooter } from '@/features/home/components/marketing-footer';
 import { useInView } from '@/hooks/use-in-view';
 import { CheckCircle2, ArrowRight, Zap } from 'lucide-react';
-import { Card } from '@/components/ui/card';
-import { Button } from '@/components/ui/button';
-import { Input, inputVariants } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
+import { Card } from '@/components/shadcn/card';
+import { Button } from '@/components/shadcn/button';
+import { Input, inputVariants } from '@/components/shadcn/input';
+import { Label } from '@/components/shadcn/label';
 import { cn } from '@/lib/utils/index';
 
 const INCLUDED = [
