+++ app/profile/edit/page.tsx
@@ -3,8 +3,8 @@
 import { useRouter } from 'next/navigation';
 import { useUser } from '@/hooks/user/use-user';
 import { ProfileEditForm } from '@/features/user/components/profile-edit-form';
-import { Button } from '@/components/ui/button';
-import { Card, CardContent } from '@/components/ui/card';
+import { Button } from '@/components/shadcn/button';
+import { Card, CardContent } from '@/components/shadcn/card';
 import { ArrowLeft } from 'lucide-react';
 import { AppLayout } from '@/features/common/components/app-layout';
 import { FloatingChat } from '@/features/chat/components/floating';
