+++ app/register/page.tsx
@@ -14,8 +14,8 @@ import {
   ShieldCheck,
   Zap,
 } from 'lucide-react';
-import { Card } from '@/components/ui/card';
-import { Badge } from '@/components/ui/badge';
+import { Card } from '@/components/shadcn/card';
+import { Badge } from '@/components/shadcn/badge';
 
 const HIGHLIGHTS = [
   {
